+++ pandas/core/reshape/api.py
@@ -1,5 +1,8 @@
+# flake8: noqa
+
 from pandas.core.reshape.concat import concat
 from pandas.core.reshape.reshape import melt
-from pandas.core.reshape.merge import merge, ordered_merge, merge_ordered, merge_asof
+from pandas.core.reshape.merge import (
+    merge, ordered_merge, merge_ordered, merge_asof)
 from pandas.core.reshape.pivot import pivot_table, crosstab
 from pandas.core.reshape.tile import cut, qcut
